chore(PlayerModule): reorganize file structure
