+++ src/app/player/components/player-controls.scss
