+++ src/app/player/components/player-controls.spec.ts
