+++ src/app/player/components/player.scss
