+++ src/app/player/components/player.spec.ts
