+++ src/app/player/components/player.ts
