+++ src/app/player/index.ts
@@ -1,0 +1,3 @@
+export { playerReducer, PlayerState, timesReducer, TimesState, TimesStateRecord } from './state';
+export { PlayerModule } from './player-module';
+export { PlayerService } from './player-service';
