+++ src/app/player/player-module.ts
@@ -1,23 +1,19 @@
 import { NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
 
-import { SharedModule } from 'src/app/shared';
-
+// components
 import { PlayerComponent } from './components/player';
 import { PlayerControlsComponent } from './components/player-controls';
 import { FormatVolumePipe } from './pipes/format-volume';
 
+// modules
+import { SharedModule } from 'src/app/shared';
+
+// services
 import { AUDIO_SOURCE_PROVIDER } from './audio-source';
 import { PlayerActions } from './player-actions';
 import { PlayerEffects } from './player-effects';
 import { PlayerService } from './player-service';
-
-
-export { PlayerService };
-export { playerReducer } from './reducers/player-reducer';
-export { PlayerState } from './reducers/player-state';
-export { timesReducer } from './reducers/times-reducer';
-export { TimesState, TimesStateRecord } from './reducers/times-state';
 
 
 @NgModule({
