+++ src/app/player/player-service.ts
@@ -7,9 +7,9 @@ import { Observable } from 'rxjs/Observable';
 import { PLAYER_INITIAL_VOLUME } from 'src/constants';
 import { AppState } from 'src/app';
 import { Track, TracklistCursor } from 'src/tracklists';
-import { PlayerState } from './reducers/player-state';
-import { getPlayer, getPlayerTrack, getPlayerTracklistCursor, getTimes } from './reducers/selectors';
-import { TimesState } from './reducers/times-state';
+import { PlayerState } from './state/player-state';
+import { getPlayer, getPlayerTrack, getPlayerTracklistCursor, getTimes } from './state/selectors';
+import { TimesState } from './state/times-state';
 import { AudioService } from './audio-service';
 import { AudioSource } from './audio-source';
 import { PlayerActions } from './player-actions';
