+++ src/app/player/state/index.ts
@@ -1,0 +1,4 @@
+export { playerReducer } from './player-reducer';
+export { PlayerState } from './player-state';
+export { timesReducer } from './times-reducer';
+export { TimesState, TimesStateRecord } from './times-state';
