+++ src/app/player/state/player-reducer.spec.ts
