+++ src/app/player/state/player-reducer.ts
