+++ src/app/player/state/times-reducer.ts
